requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
psycopg2-binary>=2.8,<3.0
drf-yasg>=1.20.0,<2.0.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
pytz==2024.2
sqlparse==0.5.3
typing_extensions==4.12.2
